+++ CHANGELOG.rst
@@ -11,10 +11,16 @@ This project adheres to `Semantic Versioning`_ starting with version 0.7.0.
 
 Added
 -----
+
+- ``EndpointConfig`` class that handles authenticated requests (ported from Rasa Core)
+- ``DataRouter()`` class supports a ``model_server`` ``EndpointConfig``, which it regularly queries to fetch NLU models
+- this can be used with ``rasa_nlu.server`` with the ``--endpoint`` option (the key for this the model server config is ``model``)
+- docs on model fetching from a URL
 - ability to specify lookup tables in training data
 
 Changed
 -------
+- loading training data from a URL requires an instance of ``EndpointConfig``
 
 - Changed evaluate behaviour to plot two histogram bars per bin.
   Plotting confidence of right predictions in a wine-ish colour
@@ -25,6 +31,7 @@ Removed
 
 Fixed
 -----
+- re-added support for entity names with special characters in markdown format
 
 [0.13.2] - 2018-08-28
 ^^^^^^^^^^^^^^^^^^^^^